Fixed some bugs

+++ 02_setting_out/ex_07_time.cpp
@@ -1,4 +1,4 @@
-// Write a program that asks the user to enter an hour value and a minute value.The
+// Write a program that asks the user to enter an hour value and a minute value. The
 // main() function should then pass these two values to a type void function that displays the two 
 // values in the format shown in the following sample run:
 // Enter the number of hours: 9
@@ -12,7 +12,7 @@ void time(int hours, int minutes);
 
 int main()
 {
-    int hours, minutes;
+    unsigned short hours, minutes;
     cout << "Enter the number of hours: ";
     cin >> hours;
     cout << "Enter the number of minutes: ";
